make extractIdsFromSimplePattern public
