+++ @xen-orchestra/backups/extractIdsFromSimplePattern.js
