+++ packages/xo-server/src/api/_backupGuard.mjs
@@ -1,5 +1,5 @@
-import { extractIdsFromSimplePattern } from '@xen-orchestra/backups/_extractIdsFromSimplePattern.js'
 import { createPredicate } from 'value-matcher'
+import { extractIdsFromSimplePattern } from '@xen-orchestra/backups/extractIdsFromSimplePattern.js'
 import { forbiddenOperation } from 'xo-common/api-errors.js'
 
 export default async function backupGuard(poolId) {
